[#55189] Use PageHeader tabs in admin/settings/project_custom_fields/:id/edit pages.
