--- app/components/admin/settings/project_custom_fields/project_attribute_detail_header_component.html.erb
@@ -1,47 +1,0 @@
-<%#-- copyright
-OpenProject is an open source project management software.
-Copyright (C) 2012-2023 the OpenProject GmbH
-
-This program is free software; you can redistribute it and/or
-modify it under the terms of the GNU General Public License version 3.
-
-OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
-Copyright (C) 2006-2013 Jean-Philippe Lang
-Copyright (C) 2010-2013 the ChiliProject Team
-
-This program is free software; you can redistribute it and/or
-modify it under the terms of the GNU General Public License
-as published by the Free Software Foundation; either version 2
-of the License, or (at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program; if not, write to the Free Software
-Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-
-See COPYRIGHT and LICENSE files for more details.
-
-++#%>
-<%= render(Settings::ProjectCustomFields::EditFormHeaderComponent.new(custom_field: @custom_field)) %>
-
-<%=
-  render(Primer::Alpha::TabNav.new(label: nil, test_selector: :project_attribute_detail_header)) do |component|
-    component.with_tab(
-      selected: project_custom_field_edit_selected?,
-      href:  edit_admin_settings_project_custom_field_path(@custom_field)
-    ) do |tab|
-      tab.with_text { t(:label_details) }
-    end
-
-    component.with_tab(
-      selected: project_custom_field_project_mappings_selected?,
-      href: project_mappings_admin_settings_project_custom_field_path(@custom_field)
-    ) do |tab|
-      tab.with_text { I18n.t(:label_project_plural) }
-    end
-  end
-%>
