--- app/components/admin/settings/project_custom_fields/project_attribute_detail_header_component.rb
@@ -1,57 +1,0 @@
-# frozen_string_literal: true
-
-#-- copyright
-# OpenProject is an open source project management software.
-# Copyright (C) 2012-2024 the OpenProject GmbH
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License version 3.
-#
-# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
-# Copyright (C) 2006-2013 Jean-Philippe Lang
-# Copyright (C) 2010-2013 the ChiliProject Team
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-#
-# See COPYRIGHT and LICENSE files for more details.
-#++
-
-module Admin
-  module Settings
-    module ProjectCustomFields
-      class ProjectAttributeDetailHeaderComponent < ApplicationComponent # rubocop:disable OpenProject/AddPreviewForViewComponent
-        TAB_NAVS = %i[
-          project_custom_field_edit
-          project_custom_field_project_mappings
-        ].freeze
-
-        def initialize(custom_field:, selected:)
-          selected = selected.to_sym
-          raise "selected must be one of the following: #{TAB_NAVS.join(', ')}" unless TAB_NAVS.include?(selected)
-
-          super
-          @custom_field = custom_field
-          @selected = selected
-        end
-
-        TAB_NAVS.each do |tab_nav|
-          define_method(:"#{tab_nav}_selected?") do
-            @selected == tab_nav
-          end
-        end
-      end
-    end
-  end
-end
