+++ app/components/settings/project_custom_fields/edit_form_header_component.rb
@@ -29,18 +29,31 @@
 module Settings
   module ProjectCustomFields
     class EditFormHeaderComponent < ApplicationComponent
-      def initialize(custom_field:)
-        super
+      TAB_NAVS = %i[
+        project_custom_field_edit
+        project_custom_field_project_mappings
+      ].freeze
 
+      def initialize(custom_field:, selected:)
+        selected = selected.to_sym
+        raise "selected must be one of the following: #{TAB_NAVS.join(', ')}" unless TAB_NAVS.include?(selected)
+
+        super
         @custom_field = custom_field
+        @selected = selected
+      end
+
+      TAB_NAVS.each do |tab_nav|
+        define_method(:"#{tab_nav}_selected?") do
+          @selected == tab_nav
+        end
       end
 
       def breadcrumbs_items
         [{ href: admin_index_path, text: t("label_administration") },
          { href: admin_settings_project_custom_fields_path, text: t("label_project_plural") },
          { href: admin_settings_project_custom_fields_path, text: t("settings.project_attributes.heading") },
-         @custom_field.name
-        ]
+         @custom_field.name]
       end
     end
   end
